requirements.txt
python-dotenv==1.0.1
requests==2.32.3
playwright==1.47.0
